+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [8.11.3] - 2019-01-17
+
+### Fixed
+
+- **Button** when using block prop and small size, button height was smaller than it was supposed to be.
+- **Table** Align toolbar buttons and use ButtonWithIcon instead of Button.
+
 ## [8.11.2] - 2019-01-16
 
 ### Added